+++ lib/fireflutter.dart
@@ -3,3 +3,4 @@ export './src/functions.dart';
 // export 'src/chat/message.model.dart';
 export 'src/presence/presence.dart';
 export 'src/presence/user.presence.dart';
+export 'src/auth/phone/phone.service.dart';