+++ lib/src/defines.dart
@@ -1,0 +1,3 @@
+typedef ErrorCallback = void Function(dynamic e);
+typedef CodeSentCallback = void Function(String verificationId);
+typedef VoidStringCallback = void Function(String);